INFRA-2334: Optimizations using AWS CLI and streaming (#12)
This PR brings in 2 performance optimizations:

1. Use AWS CLI instead of minio if configured (via input flag) since AWS
CLI has a highly optimized multithreaded upload/download mechanism that
is lacking in minio (live testing on a 10GB file showed a 3x speedup on
AWS CLI)
2. Use streaming tar compression and decompression. Instead of
downloading and then extracting from disk stream the compressed tar
stream directly to `tar`. Since uploads and downloads don't use CPU too
much streaming (de)compression can be done with **zero** overhead in all
tested cases.

Combined the two optimizations reduced a 100s download+extract of a 10GB
cache to 30s: 70% time savings! The impact will be lower on the ~2GB
caches used by gradle but should still be substantial.

I am attaching a gist showing the largest recent (~1 week) cache blobs.
For pb-ml-services the [speedup is
40%](https://github.com/productboard/pb-ml-services/actions/runs/16273218096),
we can expect decent speedups across most of the jobs producing the big
blobs on top of the list.
Big cache blobs:
https://gist.github.com/Dawnflash/d70f59b6ad1b8058a6e1418efa20500c

-
https://github.com/productboard/pb-universe/actions/runs/16516057434/job/46707037022
- 9m39s (!) cache upload
-
https://github.com/productboard/pb-ml-services/actions/runs/16492232791/job/46629471840
- 1m24s cache upload, 40% savings
-
https://github.com/productboard/graphql-node-resolver/actions/runs/16515959858/job/46706751865
- 1m55s cache upload

I tested pseudorandom files of different sizes with the optimizations on
and off (all 4 combinations).
Testing was done using the latest linux amd64 github runner image
running on k8s under
[ARC](https://github.com/actions/actions-runner-controller) on an AWS
`m6a.large` node (2c/8g). Minio library version: 8.0.4, AWS CLI version
2.27.35, tar 1.34.
Here are the times it took to compress and upload each to a VPC-local S3
bucket.

|size |base   |streaming|cli   |cli+streaming|
|-----|-------|---------|------|-------------|
|100m |2.229s |1.686s   |1.558s|2.314s       |
|500m |9.003s |7.371s   |3.084s|3.962s       |
|1024m|23.618s|13.406s  |6.191s|5.588s       |
|2048m|40.247s|27.025s  |20.233s|11.138s      |
|4096m|82.212s|51.185s  |40.845s|21.773s      |
|8192m|159.564s|98.62s   |82.083s|43.253s      |

The speedups start at 500m and amount to 56% (500M), 76% (1G), 72% (2G),
74% (4G), 73% (8G).

```yaml
with:
  use-aws-cli: true
  use-tar-streaming: true
```

This would be a fantastic upstream contribution but I'm not sure it's
fully ready. The original solution is using an internal utility API from
actions/cache which provides platform-agnostic tar calls. I did not
reverse that solution fully so the current solution here only supports
linux which isn't ideal. Still, for us it's 100% useful.

A clean solution would implement it cleanly with respect for windows,
etc. I don't have a windows machine to test on so that's why I'm
offering it here first.

We can merge it here and then offer it upstream as well and maybe ask
for collaboration on that compatibility part? Because I'm fairly sure we
couldn't expect it to get merged in the current state.